move front-end to cassette directory
