+++ cassette/components/cover.js
