+++ cassette/styles/assets/icons.js
