+++ dat.json
@@ -1,4 +1,5 @@
 {
   "url": "dat://15d8b242dfa4c1813ba76d1e6aebb2c92a99d992f4a9d3ec699076f54682ddf6/",
-  "title": "music-visions"
+  "title": "music-visions",
+  "web_root": "/cassette"
 }
