+++ songs/introduction.txt
@@ -18,4 +18,4 @@ I made a small pull request to datradio (fixing a small bug), but the creator of
 
 ~~*    
 
-Datradio exists in an entirely new space, using entirely new technology.  I did not want to give any long-term future for it, unless that future vision was also entirely new. I think this technology will allow us to create something truly meaningful for music scenes, and for the future romances of future youth.  That is really exciting to me , and so this project might have ended up a bit sincere.  But hey, the future <em>is</em> sincere.
+Datradio exists in an entirely new space, using entirely new technology.  I did not want to give any long-term future for it, unless that future vision was also entirely new. I think this technology will allow us to create something truly meaningful for music scenes, and for the future romances of future youth.  That is really exciting to me , and so this project might have ended up a bit sincere.  But hey, the future _is_ sincere.
